Updated openstack/openstack Project: openstack/ironic-python-agent  2c0ebdea09488e10a541143e3f533f51ea90c515
Adds more functional tests for commands

This adds additional functional tests for the commands API.
It also restructures the current functional test to run in
sequence as a monolithic test to preserve IPA state and test
order between test runs. The time to wait for IPA to start
was increased due to an intermittent race condition that
occurs with the larger test suite.

Partial-Bug: 1492036
Change-Id: Iff9b41fb531d34225d702b9bfd39826e3c7195ad