Updated openstack/openstack Project: openstack/nova-specs  a538f8e2264a1b7b35e8b899c45fb56038f79a0b
Persist the RequestSpec object

The RequestSpec used for scheduling an instance should be persisted so
that it can be used for any later scheduling decisions.  Things like
resize or migrate should have access to the same scheduling constraints
as the original boot request.

This is a reproposal of a spec approved for Liberty.

blueprint persist-request-spec

Previously-Approved: Liberty

Change-Id: Iecbe60b5f0139fc4802104efb40a31e7c8a6b25d