Updated openstack/openstack Project: openstack/ironic  cee8b22a1131476946ef2a648e3678fe9609d647
To run a specific unit test with ostestr use -r

To run a specific unit test using ostestr (instead of testr), the '-r'
argument needs to be specified. This updates the developer documention
to reflect this.

By default, ostestr runs the tests in parallel, so the (incorrect)
example for running them in parallel was removed.

Change-Id: I01be19d4419155b7bb5367a14a585a1016bc799f