Updated openstack/openstack Project: openstack/nova  6b43bf8a42b7f4a2528ad257212ba0234ca6a42a
Neutron: update cells when saving info_cache

Updates to the instance info_cache are not synced up to parent cells by
default.  So when Neutron allocates a port for an instance and the
info_cache is updated that is not synced up.  This used to be alright
because the instance update that followed shortly after would sync the
info_cach.  This changed when cells moved to passing instance objects
around.  The change tracking within the object means that info_cache
isn't updated in the parent cell anymore because it's not 'dirty'.
Fixing the object to update info_cache doesn't seem like the proper fix
since that was a fortunate side effect before.  Instead we should
explicitly sync info_cache up to the parent cell when we know it's being
set.

Change-Id: I6f6f10bd414b7c3f423a8f76517765748878b7d0
Closes-bug: #1471958