Updated openstack/openstack Project: openstack/swift  75e10aa77c9edc40f94514b89a11d7dab396e314
Input validation must not depend on the locale

storage_policy.py: replace string.letters with string.ascii_letters.

This change does not change the behaviour on Python 2. It is only needed
for Python 3 because string.letters was removed.

On Python 2, string.letters is modified when locale.setlocale() is
called. Hopefully, Swift doesn't call setlocale() and so it's safe to
replace it.

Change-Id: Ifbf9332ae739b1bfc9e6d2831f4e7581e69f233d