Updated openstack/openstack Project: openstack-infra/os-loganalyze  a9b91f6a05897485614b3943fb1e9222d5fddb59
Print file sizes in indexes for on disk files

File sizes are an important piece of information in log file indexes as
they help readers pick out inappropriately sized files which may
indicate the bug can be found there. Add file sizes to the index so that
we don't lose this info when compared to mod_autoindex.

Change-Id: I248a4fc9bd5ff3fa536e30bcc981a2d5f640307c