Updated openstack/openstack Project: openstack-infra/project-config  aef075e7e45295edf47074453a708188067912b7
Run npm upload job on relase slave

We need to be explicit that we want the npm upload job to run on the
release slave otherwise gearman will attempt to run it on the default
slave label which will fail because that job won't be registered.

To do this rename the job so that it matches existing convention which
makes the regex easier to read/write. Then update the reuseable slave
regex to include the job which will force it to run on release slave
without offlining the slave.

Change-Id: I26355680f2d8fc3fa5f514370d1586deb701c21c