Updated openstack/openstack Project: openstack/neutron-vpnaas  fe8130136adb730100713205fe01d44b31b4b176
Fix argument order for assertEqual to (expected, observed)

assertEqual expects that the arguments provided to it should be
(expected, observed).If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit Tests
fail. There are several Unit Test files where the arguments for assertEqual
have been swapped. This commit fixes the issue in vpnaas.

Change-Id: Id87b6727f1f292fd9a5f3e3a79c0a5409a2f8f1e
Partial-Bug: #1259292
Related-Bug: #1259292