Updated openstack/openstack Project: openstack/heat  f5c415accfff2ad399194b30b9408a903042f009
Use LOG.error when no exception handler is used

There's no need to use LOG.exception here since sys.exc_info()
results in (None, None, None) and thus would effectively just
be LOG.error.

Also, the failure of invoking LOG.exception outside of an
exception handler is due to a bug in Python3[1]. The fix for
that is only available in Python3.5.

[1] http://bugs.python.org/issue17911

Closes-Bug: #1482261

Change-Id: I9ceac0fceafa45769cc15f56027c981182ff0750