Updated openstack/openstack Project: openstack-dev/devstack  d111c2174e2f6c19975448f0271aa0c51e714624
add big warning for extras.d usage

This adds a warning for extras.d usage. This will give us something to
keep an eye on in logstash to build up the list of projects that will
break at Mitaka-1.

This also makes the deprecated handling done through a function, which
will hopefully make it more consistent in the future.

Change-Id: Icd393bc3e7095fe58be0fd13ef74fece3aa5c5f1