Updated openstack/openstack Project: openstack/ironic-python-agent  4a248f3ea05bcc76a072fcfad1eddb8387df4b26
Use mount -t sysfs to avoid host /sys dependencies

This patch updates the _install_grub2 function in image.py
so that we use 'mount -t sysfs' instead of bind mounting
the hosts /sys into the chroot.

This resolves issues which can occur with unmounting
the chroot mounts where 'target is busy' errors
may occur when you are debugging things with a serial
console.

Change-Id: I5c9a6546d048cbf54695329d9744fda55127bd30
Closes-bug: #1503385