Updated openstack/openstack Project: openstack/ironic  28f9f1275a1c7038976f468f17ed5ee6dfb3ccce
Add retries to ssh._get_hosts_name_for_node

It is possible for ssh._get_hosts_name_for_node to return None when
the driver is not able to find a VM corresponding to given MAC
addresses or when incorrect information about nodes is returned by
SSH call to VM manager. In the latter case _get_hosts_name_for_node
should be retried.

This adds two new configs to the [ssh] group:
 - 'get_vm_name_attempts': number of attempts to try to get VM name
   corresponding to a node's MAC address
 - 'get_vm_name_retry_interval': time (secs) to wait between
   attempts to get VM name

Closes-bug: #1493748

Change-Id: I0457113d57f82cccc5cb353123fb087f6ba63aa2