Updated openstack/openstack Project: openstack/nova  cde5b5b99316d1ed016ee01db11c20899782e2f2
Add test of claim context manager abort

Removes the successful-path unit test of the instance claim context
manager out of test_resource_tracker because there is no difference in
the code paths between the use of the context manager successfully and
using the claim object directly. Adds a new unit test to the
test_tracker that verifies the claim.abort() method is called when an
exception is raised during the claim object acting as a context manager.

Change-Id: I98724e1e58dfc5320ee672f48f072472bee26f4e