Updated openstack/openstack Project: openstack-infra/shade  7a347bf7a3fec7c8bd0a76abc32756d4535ac5ee
Use keystone v3 service type argument

In v3 this changed to be just type (completely ignoring, btw, the fact
that it's a python word)

Also, recognize keystone version in post_test_hook

Check the environment variable set in the keystone v2 job and
modify the clouds.yaml file appropriately to use v2.

Also, the parameter to delete changed from id in v2 to service in v3.

Co-Authored-By: David Shrewsbury <shrewsbury.dave@gmail.com>
Change-Id: I988848b734fa0f83d149c902850db2ad119d959a