Updated openstack/openstack Project: openstack/heat-templates  bb20c86e6e4c66079bbb8d6562be068c70d7dd01
Add heat template for restarting a stack

Use mistral workflow for checking status of the given stack and
recreate stack, if VM that belongs to the stack in ERROR state or
doesn't exist. Note that it is needed to allow trusts delegation in
heat for successfull creation of a trust context.

Change-Id: I0b61177539c332db0833114e29eee7c74f2ff673