Updated openstack/openstack Project: openstack/nova  7a074293624dae6fbc39b85ab6876e80603b09a9
add docstring to virt driver interface (as-is) [1 of ?]

The virt driver interface lacks documentation about preconditions and
postconditions as well as input and output parameters. This patch set
adds documentation to the current functionality "as-is". For this I
assumed that the libvirt driver, as the most used driver, should be
the source of truth. If the libvirt driver doesn't implement a method
I used the next driver which does (mostly XEN). If the driver didn't
provide information, I used the reaction and expectation from the
compute manager class in its role as user of the driver interface.

To keep this patch set assessable I will push other documentation
enhancements with following patch sets.

Change-Id: I7cad8dcb2a9b4a74cc71353ba4da14dd77528920