Updated openstack/openstack Project: openstack/manila  06dc65b81b9ec2047fb6e19e24302ef362387892
Verify common server in Generic driver on startup

If we setup generic driver in DHSS=False mode, we need to provide
specific data for VM that will host shares and if we do not provide
such info or provide it incorrectly then we get error only trying
perform actions. But we should do validation on service start.

Change-Id: I325d6461c4b12f481769bb01f792ef850dba789c
Closes-Bug: #1502063