Updated openstack/openstack Project: openstack/neutron  aa6ec7bf44215d448b493a0eb375225f19884c58
Mock oslo policy HTTPCheck instead of urllib

We were mocking internal behavior of oslo policy by
patching urllib. This will break with the upcoming oslo
release that switches to requests.

This patch changes the mock to the HTTPCheck level and we
can leave implementation details testing up to oslo_policy.

Change-Id: I07957f01307e25f1547197c720eea6e3e7f0ef5a
Closes-Bug: #1503890