Updated openstack/openstack Project: openstack/glance_store  49a0dd770b1e017cb4fe1362b7968182389f80a1
Activate pep8 check that _ is imported

Remove the specification in tox.ini that _ is a builtin so that
it will no longer assume that _ does not need to be imported.
This helps ensure that the _ from i18n is used.

Activating this check did flag violations and they were fixed
in this patch.

Change-Id: Ia7e4c75b2126a67784683d1c22f0e24bc44113a4