Updated openstack/openstack Project: openstack/neutron  02b70525f95d3f4bfd2525bbbe7ccd22ff03619c
Only lock in set_members on mutating operations

ipset was locking on every set_members call with an external
filesystem lock. This was expensive when lots of ports that
were a part of the same security group were on the same agent.

This patch adjusts it to check if it needs to make a change before
acquiring the semaphore.

Closes-Bug: #1502930
Change-Id: I2553ab74b7d0fbada5d573246194f83d58bd7d56