Updated openstack/openstack Project: openstack/manila  80e9895f44dfcd175ed3c1b7b4e9bfb9b7dfd8f8
Make share service understand driver init failure

Recent change [1] made share manager ignoring all driver setup errors
and continue run process. So, if driver init failed and we do not have
errors in "stats reporting" then manager/driver will report that
it is OK and scheduler will consider it alive when it is not indeed.

So, add attr 'initialized' to driver base class that will be used by
share manager to get to know is it allowed to call driver methods or
not.
Manila-share service will still run, but will be marked as 'down'.

[1] I5c4c551da9d3576ea49118ef09655d2939990cb2

Change-Id: I15d1f2aaf1d7cec6e8afb93fc32e46a877261d4e
Closes-Bug: #1502809