Updated openstack/openstack Project: openstack/tempest  c625bcfb52247fc5857a99af74dcef1b0ca8f42b
Stop validating pre-provisioned credentials

Stop validating pre-provisioned credentials when requested
within the credential provider. This drops the need for
estabilishing connection to the identity service, and
reduces the amount of configuration that has to be passed
into the credentials provider for it to work.

This work is in preparation to the migration of the
pre-provisioned credentials provider to tempest-lib.

Change-Id: I825a01ff72cb3a937aafeb2104333db7113ef4d0