Updated openstack/openstack Project: openstack/requirements  0d0c719d413de78003bc09253b7d238bc1ca0f76
Adds os-win in global-requirements.txt

Hyper-V is involved in many of OpenStack components (nova, neutron,
cinder, ceilometer, etc.) and will be involved with other components
in the future.

A common library can be created in order to reduce the code duplication
between all these components (utils classes, which interacts directly
with Hyper-V through WMI), making it easier to maintain, review and
propose new changes to current and future components.

Adding os-win to the global-requirements is necessary in order to add
it as a requirement for the mentioned projects.

Partially Implements: blueprint add-os-win-library

Change-Id: I01f19821e99659bbad123533e59a25f9a3339c94