Updated openstack/openstack Project: openstack/python-cinderclient  92f3d35d5b6ba54704096288fb2a5ae93f5e1985
Adding pagination to snapshots and backups lists

Snapshot and backups pagination was recently implemented in the Cinder
backend. This patch is implementing a pagination for the snapshots and
backups on the client side in the same way that volume pagination works
using limit, marker and sort parameters.

Partial-Implements: blueprint extend-limit-implementations
Change-Id: Ie3660854407a947f7470b4dc0911704c0a31c1b4