Updated openstack/openstack Project: openstack/taskflow  c343c4689540156cfce765335703ae6cae4ccb42
Relabel internal engine 'event' -> 'outcome'

Instead of calling the variable name 'event' it is
more appropriate to call these variable names 'outcome' (or
'outcomes') since they represent the 'outcome' (EXECUTED,
or REVERTED) of a atom and are not event types.

To avoid the confusion just relabel these to be outcome(s)
instead.

Change-Id: Ia9caebe279145e4887491151d40a4f59650d40d4