Updated openstack/openstack Project: openstack/trove  5b83e892bd8236d2631b18193c5bb34599b4ca2d
Add unspecified options to backup unittests

The backup runner unittests have been intermittently failing
because some of the backup commands being returned don't include
gzip in the pipeline.

The problem is that the tests all use the same instance of the
BackupRunner and the is_zipped instance variable is being
altered by the various tests. Each test that is expecting
compression and/or encryption needs to set these variables
explicitly. If not, depending on the timing and ordering of
the tests, the variables may not be set as the test expects.

Put defaults for the variables (true) into the setup() method.
Removed code in the tests that set is_zipped or is_encrypted
to true and will rely on the defaults in setup() for these.

Change-Id: I74431aefe88e1aea15ca5160a6cb597b59d19fbd
Closes-bug: 1500199