Updated openstack/openstack Project: openstack/keystonemiddleware  a29b50dd64f7eefa5cae62990cb8f66c7c5e1bb3
Use request helpers for token_info/token_auth

There is already a request setter for token_auth that we aren't using
and should be. Also add a setter for token_info rather than set the
environment directly.

This makes it much easier to consume these values later in objects that
subclass BaseAuthProtocol.

Change-Id: I7f4bf2950f783e4dc85a5671def6a02322cd7f8a