Updated openstack/openstack Project: openstack/openstack-manuals  ed2b71e51b411cfe3dbd14cd240f14558e4f417f
Add how to use in Ironic Section

Since Ironic is an integration project now. We should add one section
about it in Admin guide.
This commit is part of the bugfix since adding the whole section is
huge. This one is to add the sequence of how to use Ironic.

Change-Id: I35176fcec960110bfe62c68251a1dfd5d0d828cf
Partial-Bug: #1495898