Updated openstack/openstack Project: openstack/os-net-config  cf68668803886fb68b0f0a4a33edf65cf17c6b7c
Add support for Linux Bonding to os-net-config ifcfg

This change adds support for Linux Bonding to the impl_ifcfg
in os-net-config. This change adds support for configuring Linux
Bonds using the Bonding module rather than Open vSwitch. Most of
the options for Linux Bonds are the same as OVS, with the exception
of bonding_options instead of ovs_options.

Change-Id: If8c6de1554234277843de9fac58536dd5b0a941b