Updated openstack/openstack Project: openstack/tempest  005ff334d485c4ca231d7ee8396d3eb979a9ce59
Handle exceptions from plugins gracefully

If a tempest plugin is poorly constructed and raises an exception
during register_opts this can interfere with the rest of tempest
running correctly. This commit makes the tempest plugin handling
a bit more defensive so things don't crash if a plugin raises an
exception, instead the error is logged and tempest moves on.

Change-Id: Ic133eee1cced5ca3c53334d3d30c7b0b043ff789