Updated openstack/openstack Project: openstack/tempest-lib  9a68f4d12583b413add200ec86a415bf2a2f0704
Set a default value for default param

This commit sets a default value for the default param in the get()
method for the Credential class. The fact that you have to set a
default breaks the assumption on a get() method where it'll default
None if there is no entry for the input parameter. This commit fixes
this by defaulting the default param to be None.

Change-Id: I4015106b76f2762d0caaf73b1525c27197f803f1