Updated openstack/openstack Project: openstack/requirements  29b59f793ee0e7c90686bee0ba778133a459455b
Temporarily cap webob<1.5.0 in g-r to unblock nova

webob 1.5.0 has a change which breaks unit tests in nova, cinder
and manila.

If04eedb402092ca2cc307cb52f2231493c4375a2 fixes the nova change but
is blocked by the oslo.messaging 2.6.0 bug 1505295 which would be
fixed in I45189bb8d31819e55e2955b22aa1146257578d14 but that is
failing because of the webob failure, so we have multiple blocking
issues in nova.

Without squashing changes that we have to backport to stable/liberty
and muck with the history, we can pin both webob and oslo.messaging
versions in nova, merge that change, then land the fix for the unit
tests in nova for the webob bug, and then remove the cap on webob
here (hence the TODO).

Change-Id: I7a2197102bd812557b3a40d0a24e9a1363dd0dd7
Partial-Bug: #1505153