Updated openstack/openstack Project: openstack/neutron  3da8a28315c6bf3e3ac347efb496e4490d37942e
Add OpenFixture and get rid of 'open' mocks

This patch adds a new fixture that allows a specific
file to be mocked for 'open' calls while leaving the
normal 'open' behavior for everything else. This is to
prevent cases like the related bug where a library we
depend on can unexpectedly call 'open' on its own
(in that case it was debtcollector->warnings->linecache).

It also replaces all open mocks in the unit tests with
this new fixture.

Related-Bug: #1503847
Change-Id: Ie9e1a01acd5cb231718efd028d30df6c8dbcd3f0