Updated openstack/openstack Project: openstack/neutron-fwaas  1e829f31cd65c5ba8a8624ae3461cb707a55bda6
Fix argument order for assertEqual to (expected, observed)

assertEqual expects that the arguments provided to it should be
(expected, observed).If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit Tests
fail. There are several Unit Test files where the arguments for assertEqual
have been swapped. This commit fixes the issue in fwaas.

Change-Id: I3ec44666be34ee6b537437b858a80a239a9279ca
Partial-Bug: #1259292
Related-Bug: #1259292