Updated openstack/openstack Project: openstack/requirements  9736328ae0b76cde177bb3da873c347caa0703f5
bump upper constraints of keystonemiddleware and keystoneclient

We were impacted by the release of requests 2.8.0. We released
new versions of these libraries, ksm at 2.4.0 and ksc at 1.8.0,
that do not include that requests version. Our Keystone tests
pull in keystoneclient and keystonemiddleware, so we need to
rely on these new versions to unblock our gate.

Change-Id: Ie1ae6cb589cfcc60392029aa9d3d350212e2f154