Updated openstack/openstack Project: openstack/tripleo-incubator  7461b01e393931e0f4cf1ff38eadb0755a49d658
jq 1.5 support

According the JSON spec, the \n character should be encoded. That's a
policy that jq 1.5 enforce.

Because of that, it's not possible anymore the generate a JSON file with
the content of the private key without dropping the \n.

Without this patch, we get the following error:
parse error: Invalid string: control characters from U+0000 through
U+001F must be escaped at line 32, column 30

Change-Id: Id4a956443ff8c99f76a5778ed5c3a53367c8e408
Depends-On: I8ab2d691b2789e7f7e9931d67e9905b34fbba56