Updated openstack/openstack Project: openstack/oslo.rootwrap  31cfdbd4076bb6556cf9612171ba43fa44475d71
Fix Python 3 support for eventlet monkey-patching

Use eventlet.green.subprocess if eventlet is used and enable eventlet
tests on Python 3.

This change adds oslo_rootwrap.subprocess which is
eventlet.green.subprocess if eventlet monkey-patching is enabled or
if the TEST_EVENTLET environment variable is set, or subprocess of
the Python standard library otherwise.

When eventlet is used (with monkey-patching or not), it's more
reliable to use eventlet.green.subprocess instead of using directly
subprocess from the Python standard library.

On Python 2, it "works" to use directly subprocess: subprocess.Popen
calls os.pipe() and os.fdopen(fd) which are both monkey-patched. On
Python 3, it doesn't work because subprocess uses os.pipe() and
io.open(fd), and the io module is *not* monkey-patched at all.

Change-Id: Ib859bebe52612b35f0f1f53aedf76222683795e7