Updated openstack/openstack Project: openstack-infra/devstack-gate  5211df0123c98eb0dc96b642671ebe4cd1864440
Use devstack subunit stream in output

Devstack is starting to emit a subunit event for whether devstack
was successful or not and how long it took to execute. This will
enable us to track devstack along with tempest in subunit2sql and also
make local visualizations of devstack easier. Eventually the data in
the devstack stream will be a bit more granular which will enable
additional analysis

Change-Id: I5c2fd8ae8c33edce73836dd0c0c5107dc73685c6
Depends-On: I07112dde996c3e2c73f5aafc9b73d33d26374633