Updated openstack/openstack Project: openstack/heat  c643810f04a4fb7496eb87d3f29b734a6ecdb43d
Raise StackValidationFailed on parameter validation

This changes the _validate methods for the JsonParam and
CommaDelimitedListParam to raise StackValidationFailed execeptions if
the parameters values fail to parse. A failure to parse will be
indicated with a ValueError exception, so that exception is caught and
StackValidationFailed is raised instead.

This brings these parameters in line with the behavior in the other
parameter type classes.

It also makes the validation failures much more helpful to users with
stacks with lots of parameters (TripleO), as the name of the parameter
that failed validation will now be logged as part of the exception.

Change-Id: Ife08272f0e95491c030e5d70499fe31d462c515c