Updated openstack/openstack Project: openstack/neutron  cecf2a5883744a4aa3ece3194919b2037d007b3f
Avoid DuplicateOptError in functional tests

Some test require to register options with oslo.Config,
but now some of them use global cfg.CONF for it.
This can cause the DuplicateOptError if two tests tried to
register the same option with different values of config.
We should use the oslo_config fixture [1]
in functional tests to avoid it.

[1] http://git.openstack.org/cgit/openstack/oslo.config/tree/oslo_config/fixture.py

Closes-bug: #1505715
Change-Id: I5c2340c0af054dfa78788cbc61dfb1fd9feea535